Refactor

+++ app/Http/Controllers/V1/CommentController.php
@@ -2,15 +2,15 @@
 
 namespace App\Http\Controllers\V1;
 
-use App\Http\Controllers\Controller;
 use App\Modules\V1\Comment\DTO\CommentContent;
 use App\Modules\V1\Comment\Repositories\CommentRepositoryInterface;
 use App\Modules\V1\Comment\Requests\CreateCommentRequest;
 use App\Modules\V1\Comment\Resources\CommentResource;
 use App\Modules\V1\Comment\Services\CommentService;
 use App\Modules\V1\User\Services\UserService;
+use App\Policies\Abilities;
 
-class CommentController extends Controller
+class CommentController extends ResponseController
 {
     public function __construct(
         private CommentService $commentService,
@@ -23,7 +23,9 @@ class CommentController extends Controller
     {
         $comments = $this->commentService->all($postId);
 
-        return response()->json(['comments' => CommentResource::collection($comments)]);
+        return $this->responseOk([
+            'comments' => CommentResource::collection($comments)
+        ]);
     }
 
     public function store(CreateCommentRequest $request, $postId)
@@ -38,7 +40,7 @@ class CommentController extends Controller
             )
         );
 
-        return response()->json([
+        return $this->responseCreated([
             'message' => "Comment created successfully",
             'comment' => new CommentResource($comment)
         ]);
@@ -48,13 +50,15 @@ class CommentController extends Controller
     {
         $comment = $this->commentRepository->findByIdWithPost($postId, $commentId);
 
-        return response()->json([
+        return $this->responseOk([
             'comment' => new CommentResource($comment)
         ]);
     }
 
     public function destroy($postId, $commentId)
     {
+        $this->authorize(Abilities::DELETE, $this->commentRepository->findById($commentId));
+
         $this->commentService->delete($postId, $commentId);
 
         return response()->noContent();
