+++ app/Http/Controllers/V1/PostController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers\V1;
 
-use App\Http\Controllers\Controller;
 use App\Modules\V1\Base\DTO\PaginateQueryParams;
 use App\Modules\V1\Base\Requests\PaginateRequest;
 use App\Modules\V1\Post\DTO\PostContent;
@@ -15,8 +14,9 @@ use App\Modules\V1\Post\Resources\PostResource;
 use App\Modules\V1\Post\Resources\PostsCollection;
 use App\Modules\V1\Post\Services\PostService;
 use App\Modules\V1\User\Services\UserService;
+use App\Policies\Abilities;
 
-class PostController extends Controller
+class PostController extends ResponseController
 {
     public function __construct(
         private PostService $postService,
@@ -38,9 +38,9 @@ class PostController extends Controller
 
         $posts = $this->postService->list($paginateDto, $postListDto);
 
-        return response()->json([
-            'data' => new PostsCollection($posts)
-        ]);
+        return $this->responseOk(
+            ['data' => new PostsCollection($posts)]
+        );
     }
 
     public function feed(PaginateRequest $request)
@@ -52,7 +52,7 @@ class PostController extends Controller
 
         $posts = $this->postService->feed($paginateDto);
 
-        return response()->json([
+        return $this->responseOk([
             'data' => new PostsCollection($posts)
         ]);
     }
@@ -61,7 +61,7 @@ class PostController extends Controller
     {
         $post = $this->postRepository->findById($postId, ['profilesLiked', 'profilesReposted']);
 
-        return response()->json([
+        return $this->responseOk([
             'post' => new PostResource($post)
         ]);
     }
@@ -77,14 +77,18 @@ class PostController extends Controller
             )
         );
 
-        return response()->json([
-            'message' => 'Post created successfully',
-            'post' => new PostResource($post)
-        ]);
+        return $this->responseCreated(
+            [
+                'message' => 'Post created successfully',
+                'post' => new PostResource($post)
+            ]
+        );
     }
 
     public function update(UpdatePostRequest $request, $postId)
     {
+        $this->authorize(Abilities::UPDATE, $this->postRepository->findById($postId));
+
         $post = $this->postService->update(
             $postId,
             new PostContent(
@@ -95,11 +99,15 @@ class PostController extends Controller
             )
         );
 
-        return response()->json(['post' => new PostResource($post)]);
+        return $this->responseOk(
+            ['post' => new PostResource($post)]
+        );
     }
 
     public function destroy($postId)
     {
+        $this->authorize(Abilities::DELETE, $this->postRepository->findById($postId));
+
         $this->postService->delete($postId);
 
         return response()->noContent();
